requirements.txt
penman==0.6.2
networkx
sacrebleu
num2words
word2number
overrides
torch==1.1.0
h5py
boto3
pytorch-pretrained-bert==0.6.2
spacy
ftfy
nltk
conllu
pyyaml
pycorenlp
penman
editdistance
bs4
lxml
subword-nmt
